feat: Add One-Hot Encoding for Sex

+++ lab4/scripts/process_data.py
@@ -1,16 +1,24 @@
-# lab4/scripts/process_data.py
 import pandas as pd
 import os
 
 data_path = 'lab4/data/titanic.csv'
 
-# Загружаем датасет
-df = pd.read_csv(data_path)
+if not os.path.exists(data_path):
+    print(f"Файл не найден по пути: {data_path}")
+else:
+    df = pd.read_csv(data_path)
 
-# Выбираем только необходимые колонки: Pclass, Sex, Age
-df_processed = df[['Pclass', 'Sex', 'Age']].copy()
+    df_processed = df[['Pclass', 'Sex', 'Age']].copy()
 
-# Перезаписываем исходный файл обработанными данными
-df_processed.to_csv(data_path, index=False)
+    # Заполняем пропуски в 'Age'
+    age_mean = df_processed['Age'].mean()
+    df_processed['Age'].fillna(age_mean, inplace=True)
+    df_processed['Age'] = df_processed['Age'].astype(int)
 
-print(f"Датасет обработан. В файле '{data_path}' оставлены только колонки Pclass, Sex, Age.")
+    # Применяем One-Hot к 'Sex'
+    df_processed = pd.get_dummies(df_processed, columns=['Sex'], drop_first=True)
+
+    # Перезаписываем
+    df_processed.to_csv(data_path, index=False)
+
+    print(f"Датасет обработан: добавлен One-Hot Encoding для 'Sex'.")
